+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add `ExZample.build_list/3`
+
 ## [0.1.0] - 2020-02-29
 ### Added
 - Add `ExZample.build/1` and the core `Behaviour` that is from `example/0`